refactor: refactor progress and modal to use named export
