--- src/components/ui/Modals/DeleteResourceModal/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./DeleteResourceModal";
-export * from "./DeleteResourceModal";
