--- src/components/ui/Modals/ModalBase/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./ModalBase";
-export * from "./ModalBase";
