+++ src/components/ui/Modals/index.ts
@@ -1,5 +1,2 @@
-import DeleteResourceModal from "./DeleteResourceModal";
-import type { DeleteResourceModalProps } from "./DeleteResourceModal";
-
-export { DeleteResourceModal };
-export type { DeleteResourceModalProps };
+export { DeleteResourceModal } from "./DeleteResourceModal";
+export type { DeleteResourceModalProps } from "./DeleteResourceModal";
